Fix for random delays on some messages.
"Immediate" messages (PINGs) were emptying the bucket without triggering
the refresh timer, this put the bucket severely-negative after a long
period of inactivity.

Eventually, when a message was sent to the bot, it would attempt to send,
realise the bucket was very low on capactiy and start refilling it. The
bucket would eventually reach a positive capacity and the bot would
finally send the message.

This fix does 2 things:

1) "Immediate" messages trigger a bucket refresh
2) The bucket refresh timer now enforces a minimum capacity of 0